+++ Source/IGListAdapterUpdater.m
@@ -199,7 +199,7 @@ static NSArray *objectsWithDuplicateIdentifiersRemoved(NSArray<id<IGListDiffable
 
         executeCompletionBlocks(finished);
 
-        [delegate listAdapterUpdater:self didPerformBatchUpdates:updateData withCollectionView:collectionView];
+        [delegate listAdapterUpdater:self didPerformBatchUpdates:updateData collectionView:collectionView];
 
         // queue another update in case something changed during batch updates. this method will bail next runloop if
         // there are no changes